backend/requirements.txt
# Core web framework
fastapi
uvicorn
python-multipart

# Data manipulation and analysis
pandas>=1.3.0
numpy>=1.21.0

# Machine learning and time series
scikit-learn>=1.0.0
statsmodels>=0.13.0
xgboost>=1.5.0
pmdarima  # For ARIMA models

# Deep learning (required for LSTM models per assessment requirements)
numpy==1.24.3  # Fixed version for TensorFlow compatibility
tensorflow==2.14.0  # Last stable version with numpy 1.x compatibility

# Visualization
plotly>=5.0.0

# External APIs
python-binance>=1.0.0
yfinance>=0.2.0
requests>=2.27.0
lxml  # Required for yfinance HTML parsing

# Utilities
python-dotenv>=0.19.0
joblib  # For model persistence
feedparser>=6.0.0  # For RSS feeds
websockets>=10.0  # For WebSocket support
